+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 =====
 - Tweak button size in drawer and add "Settings" button
 - Show "Brightness" text in drawer
-- Add builtin "Settings" app with settings for dark/light theme, theme color
+- Save and restore "Brightness" setting
+- Add builtin "Settings" app with settings for Light/Dark Theme, Theme Color, Restart to Bootloader
 
 0.0.5
 =====